lambda_indicators/requirements.txt
boto3==1.28.85
pandas==2.1.3
psycopg2-binary==2.9.9
pandas-ta==0.3.14b0
